CNN/Model/requirements_models.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.4
Brotli==1.0.9
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.1
debugpy==1.8.11
decorator==5.1.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.1.0
facenet-pytorch==2.6.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.4.0
gdown==5.2.0
gmpy2==2.1.2
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.70.0
h5py==3.12.1
huggingface-hub==0.29.1
humanfriendly==10.0
idna==3.7
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
jax==0.6.0
jaxlib==0.6.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.5.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
libclang==18.1.1
lightning-utilities==0.12.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
narwhals==1.29.0
nest_asyncio==1.6.0
netron==8.2.4
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnx==1.17.0
onnxruntime==1.21.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.2.0
pillow_heif==0.22.0
pip==25.0.1
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.0
psutil==6.1.1
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
pywin32==307
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==13.9.4
roboflow==1.1.58
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==75.1.0
six==1.17.0
sounddevice==0.5.1
soupsieve==2.6
stack_data==0.6.3
sympy==1.13.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow_intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
torch==2.2.2
torchaudio==2.2.2
torchinfo==1.8.0
torchmetrics==1.6.1
torchsummary==1.5.1
torchview==0.2.6
torchvision==0.17.2
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.76
ultralytics-thop==2.0.14
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.44.0
widgetsnbextension==4.0.13
win-inet-pton==1.1.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
